core: rm unnecessary InternalNameResolverProvider
